+++ routing/route1.config
@@ -1,0 +1,38 @@
+Number of Nodes:
+9
+Next Node:
+A
+Connected Nodes:
+[(B,N);(F,E)]
+Next Node:
+B
+Connected Nodes:
+[(A,S);(E,E);(C,N)]
+Next Node:
+C
+Connected Nodes:
+[(B,S);(D,E)]
+Next Node:
+D
+Connected Nodes:
+[(C,W);(E,S);(I,E)]
+Next Node:
+E
+Connected Nodes:
+[(B,W);(D,N);(H,E);(F,S)]
+Next Node:
+F
+Connected Nodes:
+[(A,W);(E,N);(G,E)]
+Next Node:
+G
+Connected Nodes:
+[(F,W);(H,N)]
+Next Node:
+H
+Connected Nodes:
+[(E,W);(G,S);(I,N)]
+Next Node:
+I
+Connected Nodes:
+[(D,E);(H,S)]
